Moved readToken() into github.js
